// eUprava-Frontend/src/app/models/prelaz.ts
export interface Prelaz {
    id?: string;
    datum?: Date;
    imePutnika?: string;
    prezimePutnika?: string;
    JMBGPutnika?: string;
    brojLicneKartePutnika?: string;
    brojPasosaPutnika?: string;
    drzavljanstvoPutnika?: string;
    markaVozila?: string;
    modelVozila?: string;
    svrhaPutovanja?: string;
    odobren?: boolean;
  }